修改项目 添加jwt验证 修改部分代码 添加权限角色

+++ src/main/java/org/demo/studentscore/model/entity/Major.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @TableName("major")
 public class Major {
-    private Integer mid;
+    private Long mid;
     private String name;
-    private Integer majorsId;
+    private Long majorsId;
 }
